requirements.txt
python-dotenv
coinbase-advanced-py
flask
gunicorn